refactor(server-action) : restructure APIs
- rename folder to server-action
- separate logic based on route
