--- src/api/api.backend.ts
@@ -1,645 +1,0 @@
-'use server';
-
-import { MCQ } from '@/model/interface/MCQ';
-import { CuratedArtWorkAttribute } from '@/model/interface/curatedArtwork-types';
-import { serverLogger } from '@/util/logger';
-import { Painting, Style } from '../model/interface/painting';
-import {
-    CreateOneTimeTokenDTO,
-    CreateQuizDTO,
-    CreateUserDTO,
-    FindPaintingResult,
-    FindQuizResult,
-    QuizContextDTO,
-    QuizReactionDTO,
-    ReplacePassWordDTO,
-    ReplaceUsernameDTO,
-    requestVerificationDTO,
-    ResponseQuizDTO,
-    SendOneTimeTokenDTO,
-    SignInResponse,
-    VerifyDTO,
-} from './dto';
-import { Quiz, QuizDislike, QuizLike } from '../model/interface/quiz';
-import { User } from '../model/interface/user';
-import {
-    ENUM_ONE_TIME_TOKEN_HEADER,
-    ENUM_SECURITY_TOKEN_HEADER,
-    QuizReactionType,
-    QuizSubmitDTO,
-} from './api.backend.option';
-import { RequestQueryBuilder } from '@dataui/crud-request';
-import { Tag } from '../model/interface/tag';
-
-// TODO: HTTP API 에러 핸들링 로직 추가
-// - [ ] : fetch()가 반환한 응답 상태 확인 및 에러 핸들링 로직 추가
-// - [ ] : fetch() 동작 도중 발생하는 에러 핸들링 로직 추가
-// - [ ] : fetch() 오류 발생시, front-end의 요청을 추적하기 위해서 request-id 추가하기
-//  -> 또는 backend에서 반환한 응답에 백엔드의 request ID 삽입하기.
-//  -> <할 일 > 설명 ( 생략가능 )
-// - [ ] <추가 작업>
-// ! 주의: <경고할 사항>
-// ? 질문: <의문점 또는 개선 방향>
-// * 참고: <관련 정보나 링크>
-
-function getServerUrl(): string {
-    const url = process.env.BACKEND_URL;
-    serverLogger.info(`BACKEND_URL=${url} `);
-
-    if (url == undefined) {
-        console.error(` 'process.env.BACKEND_URL' not read`);
-        return '';
-    }
-    return url;
-}
-// TODO page.tsx 최소화 예정 (데이터 처리 함수 옮길 예정)
-export const getWeekArtWorkData = async (): Promise<CuratedArtWorkAttribute[]> => {
-    // const response = await fetch('http://localhost:4000/api/artwork_week', {
-    //     cache: 'no-cache',
-    // });  // src/data에 파일을 읽어 올 때 사용
-    const serverUrl = getServerUrl();
-
-    const url: string = serverUrl + '/painting/artwork_of_week';
-    const response = await fetch(url, {
-        cache: 'no-cache',
-    }); // 서버에 있는 데이터 읽어 올때 사용
-    const res = await response.json();
-    return res.data;
-};
-
-export const getMCQData = async (): Promise<MCQ[]> => {
-    // const response = await fetch('http://localhost:4000/api/mcq', {
-    //     cache: 'no-cache',
-    // });  // src/data에 파일을 읽어 올 때 사용
-    const serverUrl = getServerUrl();
-
-    const url: string = serverUrl + '/quiz/quiz_of_week';
-    const response = await fetch(url, {
-        cache: 'no-cache',
-    }); // 서버에 있는 데이터 읽어 올때 사용
-    const res = await response.json();
-    return res.data;
-};
-
-export const findPainting = async (
-    title: string = '',
-    artist: string = '',
-    tags: string[] = [],
-    styles: string[] = [],
-    page: number = 0,
-): Promise<FindPaintingResult> => {
-    const backendUrl = getServerUrl();
-    const titleParam = `title=${title}`;
-    const artistParam = `artistName=${artist}`;
-    const tagParam = tags.map((t) => `tags[]=${t}`).join('&');
-    const styleParam = styles.map((s) => `styles[]=${s}`).join('&');
-    const url = `${backendUrl}/painting?${titleParam}&${artistParam}&${tagParam}&${styleParam}&page=${page}`;
-    serverLogger.info(`[findPaintings] url=${url}`);
-    const response = await fetch(url);
-    const result: FindPaintingResult = await response.json();
-    return result;
-};
-
-export const getPainting = async (id: string): Promise<Painting | undefined> => {
-    try {
-        const backendUrl = getServerUrl();
-        const url = `${backendUrl}/painting/by-ids?ids[]=${id}`;
-        const response = await fetch(url);
-        const paintings: Painting[] = await response.json();
-
-        if (response.ok) {
-            if (paintings.length === 0) {
-                return undefined;
-            }
-
-            return paintings.at(0);
-        }
-
-        if (response.status === 400) {
-            return undefined;
-        }
-
-        throw Error(
-            `error from backend. status : ${response.status}, statusText : ${response.statusText}`,
-        );
-    } catch (e: unknown) {
-        serverLogger.error(`[getPainting] error : ${JSON.stringify(e)}`);
-        throw e;
-    }
-};
-
-export const findQuiz = async (
-    artists: string[] = [],
-    tags: string[] = [],
-    styles: string[] = [],
-    page: number = 0,
-): Promise<FindQuizResult> => {
-    const backendUrl = getServerUrl();
-    const artistsParam = artists.map((a) => `artists[]=${a}`).join('&');
-    const tagParam = tags.map((t) => `tags[]=${t}`).join('&');
-    const styleParam = styles.map((s) => `styles[]=${s}`).join('&');
-    const url = `${backendUrl}/quiz?${artistsParam}&${tagParam}&${styleParam}&page=${page}`;
-    serverLogger.info(`[findQuiz] url=${url}`);
-    const response = await fetch(url);
-    const result: FindQuizResult = await response.json();
-    return result;
-};
-
-export const getQuiz = async (id: string): Promise<Quiz> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/${id}`;
-    serverLogger.info(`[getQuiz] url=${url}`);
-    const response = await fetch(url);
-    const result: Quiz = await response.json();
-    serverLogger.debug(`[getQuiz] ${JSON.stringify(result, null, 2)}`);
-    return result;
-};
-
-export const addQuiz = async (jwt: string, dto: CreateQuizDTO): Promise<Quiz | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz`;
-    const response = await fetch(url, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${jwt}`,
-        },
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`add Quiz fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-    const result: Quiz = await response.json();
-
-    return result;
-};
-
-export const submitQuiz = async (
-    quizID: string,
-    dto: QuizSubmitDTO,
-): Promise<boolean | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/submit/${quizID}`;
-
-    const response = await fetch(url, {
-        method: 'POST',
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`signUp fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: boolean = await response.json();
-    return result;
-};
-
-export const getQuizReactions = async (
-    quizID: string,
-    type: QuizReactionType,
-    page?: number,
-    findUserId?: string,
-): Promise<QuizDislike[] | QuizLike[] | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/${quizID}/reactions`;
-    const typeParam = `type=${type}`;
-    const pageParam = `page=${page}`;
-    const userIdParam = `user_id=${findUserId}`;
-    const response = await fetch(url + `?${typeParam}&${pageParam}&${userIdParam}`);
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`getQuizReactions fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: QuizDislike[] | QuizLike[] = await response.json();
-    return result;
-};
-
-export const addQuizReactions = async (
-    jwt: string,
-    quizID: string,
-    dto: QuizReactionDTO,
-): Promise<boolean | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/${quizID}/reactions`;
-    const headers = {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${jwt}`,
-    };
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`getQuizReactions fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    return true;
-};
-
-export const deleteQuizReaction = async (
-    jwt: string,
-    quizID: string,
-): Promise<boolean | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/${quizID}/reactions`;
-    const headers = {
-        Authorization: `Bearer ${jwt}`,
-    };
-    const response = await fetch(url, {
-        method: 'DELETE',
-        headers,
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`getQuizReactions fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    return true;
-};
-
-export const scheduleQuiz = async (
-    context?: QuizContextDTO,
-    currentIndex?: number,
-    endIndex?: number,
-): Promise<ResponseQuizDTO | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/schedule`;
-    const currentIndexParam = `currentIndex=${currentIndex}`;
-    const endIndexParam = `endIndex=${endIndex}`;
-    const contextParam = `context=` + JSON.stringify(context);
-    const response = await fetch(url + `?${currentIndexParam}&${endIndexParam}&${contextParam}`);
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`getQuizReactions fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-    const result: ResponseQuizDTO = await response.json();
-    return result;
-};
-
-export const addQuizContext = async (dto: QuizContextDTO): Promise<boolean | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/quiz/schedule`;
-    const response = await fetch(url, {
-        method: 'POST',
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`getQuizReactions fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-    const result: boolean = await response.json();
-    return result;
-};
-
-export const signUp = async (
-    jwt: string,
-    jwtID: string,
-    dto: CreateUserDTO,
-): Promise<User | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user`;
-
-    const headers = {
-        'Content-Type': 'application/json',
-        [ENUM_ONE_TIME_TOKEN_HEADER.X_ONE_TIME_TOKEN]: jwt,
-        [ENUM_ONE_TIME_TOKEN_HEADER.X_ONE_TIME_TOKEN_ID]: jwtID,
-    };
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`signUp fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: User = await response.json();
-    return result;
-};
-
-export const getUser = async (id: string): Promise<User | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user/`;
-    const response = await fetch(url + `/${id}`);
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`getUser fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: User = await response.json();
-    return result;
-};
-
-export const findUsers = async (queryBuilder: RequestQueryBuilder): Promise<User[] | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user/`;
-    const response = await fetch(url + `?${queryBuilder.query()}`);
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`findUsers fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: User[] = await response.json();
-    return result;
-};
-
-export const updateUserPW = async (
-    user: User,
-    securityJWT: string,
-    securityJWTId: string,
-    dto: ReplacePassWordDTO,
-): Promise<boolean> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user/${user.email}/password`;
-    const headers = {
-        'Content-Type': 'application/json',
-        [ENUM_SECURITY_TOKEN_HEADER.X_SECURITY_TOKEN_ID]: securityJWT,
-        [ENUM_SECURITY_TOKEN_HEADER.X_SECURITY_TOKEN]: securityJWTId,
-    };
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`updateUserPW fail. ${JSON.stringify(result, null, 2)}`);
-        return false;
-    }
-
-    return true;
-};
-
-export const updateUserUsername = async (
-    user: User,
-    jwt: string,
-    dto: ReplaceUsernameDTO,
-): Promise<boolean> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user/${user.email}/username`;
-    const headers: HeadersInit = {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${jwt}`,
-    };
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`updateUserUsername fail. ${JSON.stringify(result, null, 2)}`);
-        return false;
-    }
-
-    return true;
-};
-
-export const deleteUser = async (
-    user: User,
-    securityJWT: string,
-    securityJWTId: string,
-): Promise<boolean> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user/${user.email}`;
-    const headers = {
-        'Content-Type': 'application/json',
-        [ENUM_SECURITY_TOKEN_HEADER.X_SECURITY_TOKEN_ID]: securityJWT,
-        [ENUM_SECURITY_TOKEN_HEADER.X_SECURITY_TOKEN]: securityJWTId,
-    };
-    const response = await fetch(url, {
-        method: 'DELETE',
-        headers,
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`deleteUser fail. ${JSON.stringify(result, null, 2)}`);
-        return false;
-    }
-
-    return true;
-};
-
-export const recoverUser = async (email: string, securityJWT: string, securityJWTId: string) => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/user/recover/${email}`;
-    const headers = {
-        'Content-Type': 'application/json',
-        [ENUM_SECURITY_TOKEN_HEADER.X_SECURITY_TOKEN_ID]: securityJWT,
-        [ENUM_SECURITY_TOKEN_HEADER.X_SECURITY_TOKEN]: securityJWTId,
-    };
-    const response = await fetch(url, {
-        method: 'PATCH',
-        headers,
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`updateUserPW fail. ${JSON.stringify(result, null, 2)}`);
-        return false;
-    }
-
-    return true;
-};
-
-export const findTags = async (queryBuilder: RequestQueryBuilder): Promise<Tag[] | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/tag/`;
-    const response = await fetch(url + `?${queryBuilder.query()}`);
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`findTags fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: Tag[] = await response.json();
-    return result;
-};
-
-export const findStyles = async (
-    queryBuilder: RequestQueryBuilder,
-): Promise<Style[] | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/style/`;
-    const response = await fetch(url + `?${queryBuilder.query()}`);
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`findStyles fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: Style[] = await response.json();
-    return result;
-};
-export const OneTimeTokenPurposeValues = {
-    UPDATE_PASSWORD: 'update-password',
-    DELETE_ACCOUNT: 'delete-account',
-    MAGIC_LOGIN: 'magic-login',
-    EMAIL_VERIFICATION: 'email-verification',
-    RECOVER_ACCOUNT: 'recover-account',
-    // SET_USER_ACTIVE : 'set-user-active',
-    // RESET_PASSWORD: 'reset-password',
-} as const;
-export type OneTimeTokenPurpose =
-    (typeof OneTimeTokenPurposeValues)[keyof typeof OneTimeTokenPurposeValues];
-
-export interface OneTimeToken {
-    id: string;
-    email: string;
-    token: string;
-    used_date: Date;
-    expired_date: Date;
-    purpose: OneTimeTokenPurpose;
-}
-
-export const signIn = async (id: string, password: string): Promise<SignInResponse | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/auth/sign-in`;
-    const credentials = btoa(`${id}:${password}`);
-    const headers = {
-        Authorization: `Basic ${credentials}`,
-        'Content-Type': 'application/json',
-    };
-
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`signIn fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: SignInResponse = await response.json();
-    return result;
-};
-
-export const requestVerification = async (
-    dto: requestVerificationDTO,
-): Promise<boolean | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/auth/request-verification`;
-    const headers = {
-        'Content-Type': 'application/json',
-    };
-
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`requestVerification fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    return true;
-};
-
-export const verifyEmail = async (dto: VerifyDTO): Promise<OneTimeToken | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/auth/verify`;
-    const headers = {
-        'Content-Type': 'application/json',
-    };
-
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`verifyEmail fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: OneTimeToken = await response.json();
-    return result;
-};
-
-export const generateSecurityToken = async (
-    id: string,
-    password: string,
-    dto: CreateOneTimeTokenDTO,
-): Promise<OneTimeToken | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/auth/security-token`;
-    const credentials = btoa(`${id}:${password}`);
-    const headers = {
-        Authorization: `Basic ${credentials}`,
-        'Content-Type': 'application/json',
-    };
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`generateSecurityToken fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    const result: OneTimeToken = await response.json();
-    return result;
-};
-
-export const sendSecurityTokenToEmail = async (
-    dto: SendOneTimeTokenDTO,
-): Promise<boolean | undefined> => {
-    const backendUrl = getServerUrl();
-    const url = `${backendUrl}/auth/security-token/send`;
-    const headers = {
-        'Content-Type': 'application/json',
-    };
-    const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(dto),
-    });
-
-    if (!response.ok) {
-        const result = await response.json();
-        serverLogger.error(`sendSecurityTokenToEmail fail. ${JSON.stringify(result, null, 2)}`);
-        return undefined;
-    }
-
-    return true;
-};
