+++ src/app/lib/api.quiz.ts
@@ -2,9 +2,9 @@
 
 import { randomInt } from 'crypto';
 import { Quiz } from '../../model/interface/quiz';
-import { findQuiz } from '../../api/api.backend';
+import { findQuiz } from '../../server-action/api.backend';
 import { QuizContext, SingletonQuizContextScheduler } from './api.quiz.scheduler';
-import { FindQuizResult } from '../../api/dto';
+import { FindQuizResult } from '../../server-action/dto';
 import { serverLogger } from '../../util/logger';
 import { Painting } from '../../model/interface/painting';
