+++ src/server-action/backend/painting/dto.ts
@@ -1,0 +1,4 @@
+import { IPaginationResult } from '..';
+import { ShortPainting } from '../../../model/interface/painting';
+
+export type FindPaintingResult = IPaginationResult<ShortPainting>;
