+++ src/server-action/backend/painting/index.ts
@@ -1,0 +1,2 @@
+export * as api from './api';
+export * as dto from './dto';
