+++ src/server-action/backend/quiz/index.ts
@@ -1,0 +1,3 @@
+export * as api from './api';
+export * as dto from './dto';
+export * as type from './type';
