+++ src/server-action/backend/quiz/type.ts
@@ -1,0 +1,6 @@
+export const QuizReactionTypeValues = {
+    LIKE: 'like',
+    DISLIKE: 'dislike',
+} as const;
+
+export type QuizReactionType = (typeof QuizReactionTypeValues)[keyof typeof QuizReactionTypeValues];
